requirements.txt
pandas
ta
binance
python-telegram-bot
